use filename comms_thread.py
